chore: optimzed code

+++ backend/entity/nodes.go
@@ -3,8 +3,8 @@ package entity
 import "database/sql"
 
 type Nodes struct {
-	ID       int64         `db:"id"`
-	ParentID sql.NullInt64 `db:"parent_id"`
+	ID       uint          `db:"id"`
+	ParentID sql.NullInt32 `db:"parent_id"`
 	Type     string        `db:"type"`
 	Name     string        `db:"name"`
 }
